Moved common SQLGetData functions to the statement_functions for reuse purposes.
