+++ tools/odbc/include/statement_functions.hpp
@@ -11,5 +11,8 @@ SQLRETURN PrepareStmt(SQLHSTMT statement_handle, SQLCHAR *statement_text, SQLINT
 
 SQLRETURN ExecuteStmt(SQLHSTMT statement_handle);
 
+SQLRETURN GetDataStmtResult(SQLHSTMT statement_handle, SQLUSMALLINT col_or_param_num, SQLSMALLINT target_type,
+                            SQLPOINTER target_value_ptr, SQLLEN buffer_length, SQLLEN *str_len_or_ind_ptr);
+
 } // namespace duckdb
 #endif
